+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to the Vz Keymap extension will be documented in this file.
 
 ### [Unreleased]
-- Changed:
-  - Changed the behavior of Paste of a text which is from box-selection mode to make more compatible with Vz Editor.
+- Improved:
+  - The behavior of Paste of a text which is from box-selection mode to be more compatible with Vz Editor.
 
 ### [0.10.0] - 2020-04-20
 - Added: