+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Vz Keymap extension will be documented in this file.
 
-### [Unreleased]
+### [0.10.1] - 2020-04-29
 - Improved:
   - The behavior of Paste of a text which is from box-selection mode to be more compatible with Vz Editor.
   - The behavior of Delete/Backspace in box-selection mode.